runtime: move payment to main mod
